Use CustomizedPairingHistoryRepositoryImpl to retrieve pairing history. Use Testcontainers to test that we are integrating with the database correctly.
Before this change, we use JPA to automagically implement this function.
That resulted in lots of unnecessary inner join queries from pairing history people table to person table. The number of queries that JPA issued for each request was around (number of pairing board in each pairing history) * (number of pairing history in the last 30 days)

After this change, when retrieving pairing history, there should always be exactly 3 queries:
1. retrieve the project
2. retrieve the pairing history for that project
3. retrieve all the people mentioned in the pairing history

So the number of queries for each request no longer grows with the number of pairing boards and/or length of history.